+++ assets/src/hooks.js
@@ -16,6 +16,7 @@ import FormHook from './hooks/Form'
 import ListingHook from './hooks/Listing'
 import LivePreviewHook from './hooks/LivePreview'
 import MapURLParserHook from './hooks/MapURLParser'
+import ModalHook from './hooks/Modal'
 import MuxUploaderHook from './hooks/MuxUploader'
 import BunnyUploaderHook from './hooks/BunnyUploader'
 import CloudflareUploaderHook from './hooks/CloudflareUploader'
@@ -63,6 +64,7 @@ export default (app) => {
     'Brando.Listing': ListingHook(app),
     'Brando.LivePreview': LivePreviewHook(app),
     'Brando.MapURLParser': MapURLParserHook(app),
+    'Brando.Modal': ModalHook(app),
     'Brando.MuxUploader': MuxUploaderHook(app),
     'Brando.BunnyUploader': BunnyUploaderHook(app),
     'Brando.CloudflareUploader': CloudflareUploaderHook(app),